api/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
pydantic==2.12.5
python-dotenv==1.2.2
requests==2.32.5
watchdog==6.0.0
firebase-admin==6.5.0
google-genai==1.70.0
